requirements.txt
streamlit
torch
torchvision
pillow
pandas
matplotlib
opencv-python-headless
Pillow
